+++ modules/editor/vs/config.el
@@ -10,7 +10,7 @@
   :init
   (setq vsc-edit-insert-tab-on-tab t)
 
-  (message-clean-mode-add-echo-commands
+  (msg-clean-add-echo-commands
    '( vsc-edit-beginning-of-line vsc-edit-end-of-line))
 
   (eldoc-add-command