+++ modules/emacs/compile/config.el
@@ -5,7 +5,7 @@
   (setq compilation-context-lines t
         compilation-scroll-output t)
 
-  (message-clean-mode-add-echo-commands '( compilation-handle-exit))
+  (msg-clean-add-echo-commands '( compilation-handle-exit))
   :config
   (require 'ansi-color)
   (jcs-add-hook 'compilation-filter-hook