+++ modules/tools/lookup/config.el
@@ -9,7 +9,7 @@
   (setq define-it-output-choice (if elenv-graphic-p 'frame 'view)
         define-it-text-scale-level -2)
 
-  (message-clean-mode-add-echo-commands '( define-it)))
+  (msg-clean-add-echo-commands '( define-it)))
 
 (use-package preview-it
   :init